+++ examples/ani1x_training.py
@@ -13,23 +13,30 @@ https://doi.org/10.1038/s41597-020-0473-z
 """
 
 import argparse
-import numpy as np
 import torch
 import hippynn
 import ase.units
 
 
-def make_model(network_params, tensor_order, atomization_consistent):
+def make_model(network_params, tensor_model, tensor_order, tensor_factors, atomization_consistent):
     """
     Build the model graph for energy and potentially force prediction.
     """
     from hippynn.graphs import inputs, networks, targets, physics
 
-    net_class={
-            0:networks.Hipnn,
-            1:networks.HipnnVec,
-            2:networks.HipnnQuad,
+    if tensor_order == 0 or tensor_model == "NONE":
+        net_class = networks.Hipnn
+    else:
+        if tensor_model == "TS":
+            net_class = {
+                1: networks.HipnnVec,
+                2: networks.HipnnQuad,
             }[tensor_order]
+        elif tensor_model == "HOP":
+            net_class = networks.HipHopnn
+            network_params["l_max"] = tensor_order
+            network_params["n_max"] = tensor_factors
+
     species = inputs.SpeciesNode(db_name="atomic_numbers")
     positions = inputs.PositionsNode(db_name="coordinates")
     network = net_class("hipnn_model", (species, positions), module_kwargs=network_params)
@@ -63,12 +70,12 @@ def make_loss(henergy, force, force_training):
         "F-RSQ": Rsq.of_node(force),
     }
 
-    losses["EnergyTotal"] = losses['T-RMSE'] + losses["T-MAE"] 
+    losses["EnergyTotal"] = losses["T-RMSE"] + losses["T-MAE"]
     losses["LossTotal"] = losses["EnergyTotal"] + losses["T-Hier"]
     if force_training:
-       losses.update(force_losses)
-       losses["ForceTotal"] = losses["F-RMSE"] + losses["F-MAE"]
-       losses["LossTotal"] = losses["LossTotal"] + losses["ForceTotal"]
+        losses.update(force_losses)
+        losses["ForceTotal"] = losses["F-RMSE"] + losses["F-MAE"]
+        losses["LossTotal"] = losses["LossTotal"] + losses["ForceTotal"]
 
     return losses
 
@@ -77,8 +84,8 @@ def make_loss(henergy, force, force_training):
 # # # Old values with singlet/triplet multiplicity only
 # # SELF_ENERGY_APPROX = {'C': -37.764142, 'H': -0.4993212, 'N': -54.4628753, 'O': -74.940046}
 # Recalculated with appropriate vacuum multiplicity
-SELF_ENERGY_APPROX = {'C': -37.8338334397, 'H': -0.499321232710, 'N': -54.5732824628, 'O': -75.0424519384}
-SELF_ENERGY_APPROX = {k: SELF_ENERGY_APPROX[v] for k, v in zip([6, 1, 7, 8], 'CHNO')}
+SELF_ENERGY_APPROX = {"C": -37.8338334397, "H": -0.499321232710, "N": -54.5732824628, "O": -75.0424519384}
+SELF_ENERGY_APPROX = {k: SELF_ENERGY_APPROX[v] for k, v in zip([6, 1, 7, 8], "CHNO")}
 SELF_ENERGY_APPROX[0] = 0
 
 
@@ -94,35 +101,38 @@ def load_db(db_info, en_name, force_name, seed, anidata_location, n_workers, use
 
     CCX_EN_NAME = "ccsd(t)_cbs.energy"
     if use_ccx_subset:
-        db_info['targets'].append(CCX_EN_NAME)
-    database = PyAniFileDB(
-        file=anidata_location,
-        species_key='atomic_numbers',
-        seed=seed,
-        num_workers=n_workers,
-        **db_info
-    )
+        db_info["targets"].append(CCX_EN_NAME)
+    database = PyAniFileDB(file=anidata_location, species_key="atomic_numbers", seed=seed, num_workers=n_workers, **db_info)
     if en_name != CCX_EN_NAME:
         database.targets = [x for x in database.targets if x != CCX_EN_NAME]
 
     # compute (approximate) atomization energy by subtracting self energies
-    self_energy = np.vectorize(SELF_ENERGY_APPROX.__getitem__)(database.arr_dict['atomic_numbers'])
-    self_energy = self_energy.sum(axis=1)  # Add up over atoms in system.
-    database.arr_dict[en_name] = (database.arr_dict[en_name] - self_energy)
-    kcalpmol = (ase.units.kcal/ase.units.mol)
-    conversion = ase.units.Ha/kcalpmol
-    database.arr_dict[en_name] = database.arr_dict[en_name].astype(np.float32)*conversion
+
+    # Build a lookup tensor for self energies
+    max_z = max(SELF_ENERGY_APPROX.keys()) + 1  # +1 in case max Z is the last index
+    lookup_table = torch.zeros(max_z, dtype=torch.float32)
+    for z, energy in SELF_ENERGY_APPROX.items():
+        lookup_table[z] = energy
+
+    database.arr_dict["atomic_numbers"] = database.arr_dict["atomic_numbers"].long()
+
+    self_energy = lookup_table[database.arr_dict["atomic_numbers"]]
+    self_energy = self_energy.sum(dim=1)
+    database.arr_dict[en_name] = database.arr_dict[en_name] - self_energy
+    kcalpmol = ase.units.kcal / ase.units.mol
+    conversion = ase.units.Ha / kcalpmol
+    database.arr_dict[en_name] = database.arr_dict[en_name].float() * conversion
+
     if force_name in database.arr_dict:
-        database.arr_dict[force_name] = database.arr_dict[force_name]*conversion
+        database.arr_dict[force_name] = database.arr_dict[force_name] * conversion
     torch.set_default_dtype(torch.float32)
-    database.arr_dict['atomic_numbers'] = database.arr_dict['atomic_numbers'].astype(np.int64)
 
     # Drop indices where computed energy not retrieved.
     if use_ccx_subset:
         filter_name = CCX_EN_NAME
     else:
         filter_name = en_name
-    found_indices = ~np.isnan(database.arr_dict[filter_name])
+    found_indices = ~torch.isnan(database.arr_dict[filter_name])
     database.arr_dict = {k: v[found_indices] for k, v in database.arr_dict.items()}
     database.make_trainvalidtest_split(test_size=0.1, valid_size=0.1)
     return database
@@ -135,19 +145,22 @@ def setup_experiment(training_modules, device, batch_size, init_lr, patience, ma
     from hippynn.experiment.controllers import RaiseBatchSizeOnPlateau, PatienceController
 
     optimizer = torch.optim.Adam(training_modules.model.parameters(), lr=init_lr)
-    scheduler = RaiseBatchSizeOnPlateau(optimizer=optimizer,
-                                        max_batch_size=batch_size,
-                                        patience=patience,
-                                        factor=0.5,
-                                        )
+    scheduler = RaiseBatchSizeOnPlateau(
+        optimizer=optimizer,
+        max_batch_size=batch_size,
+        patience=patience,
+        factor=0.5,
+    )
 
-    controller = PatienceController(optimizer=optimizer,
-                                    scheduler=scheduler,
-                                    batch_size=batch_size,
-                                    eval_batch_size=batch_size,
-                                    max_epochs=max_epochs,
-                                    stopping_key=stopping_key,
-                                    termination_patience=2 * patience)
+    controller = PatienceController(
+        optimizer=optimizer,
+        scheduler=scheduler,
+        batch_size=batch_size,
+        eval_batch_size=batch_size,
+        max_epochs=max_epochs,
+        stopping_key=stopping_key,
+        termination_patience=2 * patience,
+    )
 
     setup_params = hippynn.experiment.SetupParams(
         controller=controller,
@@ -156,16 +169,38 @@ def setup_experiment(training_modules, device, batch_size, init_lr, patience, ma
     return setup_params
 
 
-ANI1X_DSETS_KEYS = ['hf_tz.energy', 'coordinates', 'tpno_ccsd(t)_dz.corr_energy', 'wb97x_dz.hirshfeld_charges',
-                    'wb97x_tz.mbis_charges', 'wb97x_tz.forces', 'mp2_tz.corr_energy', 'npno_ccsd(t)_tz.corr_energy',
-                    'wb97x_tz.mbis_volumes', 'wb97x_tz.energy', 'wb97x_tz.dipole', 'wb97x_tz.mbis_octupoles',
-                    'wb97x_tz.mbis_quadrupoles', 'mp2_qz.corr_energy', 'wb97x_tz.mbis_dipoles', 'wb97x_dz.cm5_charges',
-                    'path', 'atomic_numbers', 'hf_qz.energy', 'mp2_dz.corr_energy', 'wb97x_dz.dipole',
-                    'npno_ccsd(t)_dz.corr_energy', 'wb97x_dz.energy', 'hf_dz.energy', 'wb97x_dz.quadrupole',
-                    'ccsd(t)_cbs.energy', 'wb97x_dz.forces']
+ANI1X_DSETS_KEYS = [
+    "hf_tz.energy",
+    "coordinates",
+    "tpno_ccsd(t)_dz.corr_energy",
+    "wb97x_dz.hirshfeld_charges",
+    "wb97x_tz.mbis_charges",
+    "wb97x_tz.forces",
+    "mp2_tz.corr_energy",
+    "npno_ccsd(t)_tz.corr_energy",
+    "wb97x_tz.mbis_volumes",
+    "wb97x_tz.energy",
+    "wb97x_tz.dipole",
+    "wb97x_tz.mbis_octupoles",
+    "wb97x_tz.mbis_quadrupoles",
+    "mp2_qz.corr_energy",
+    "wb97x_tz.mbis_dipoles",
+    "wb97x_dz.cm5_charges",
+    "path",
+    "atomic_numbers",
+    "hf_qz.energy",
+    "mp2_dz.corr_energy",
+    "wb97x_dz.dipole",
+    "npno_ccsd(t)_dz.corr_energy",
+    "wb97x_dz.energy",
+    "hf_dz.energy",
+    "wb97x_dz.quadrupole",
+    "ccsd(t)_cbs.energy",
+    "wb97x_dz.forces",
+]
 
-AVAIL_METHODS = ['hf', 'wb97x', 'ccsd(t)', 'mp2']
-AVAIL_BASIS = ['dz', 'tz', 'qz', 'cbs']
+AVAIL_METHODS = ["hf", "wb97x", "ccsd(t)", "mp2"]
+AVAIL_BASIS = ["dz", "tz", "qz", "cbs"]
 
 
 def get_data_names(qm_method, basis_set):
@@ -188,12 +223,12 @@ def main(args):
 
     hippynn.settings.WARN_LOW_DISTANCES = False
     if args.noprogress:
-        hippynn.settings.PROGRESS = None 
+        hippynn.settings.PROGRESS = None
 
     netname = f"{args.tag}_GPU{args.gpu}"
     network_parameters = {
         "possible_species": [0, 1, 6, 7, 8],
-        'n_features': args.n_features,
+        "n_features": args.n_features,
         "n_sensitivities": args.n_sensitivities,
         "dist_soft_min": args.lower_cutoff,
         "dist_soft_max": args.cutoff_distance - 1,
@@ -203,10 +238,14 @@ def main(args):
     }
 
     with hippynn.tools.active_directory(netname):
-        with hippynn.tools.log_terminal("training_log.txt", 'wt'):
-            henergy, force = make_model(network_parameters,
-                                        tensor_order=args.tensor_order,
-                                        atomization_consistent=args.atomization_consistent)
+        with hippynn.tools.log_terminal("training_log.txt", "wt"):
+            henergy, force = make_model(
+                network_parameters,
+                tensor_model=args.tensor_model,
+                tensor_order=args.tensor_order,
+                tensor_factors=args.tensor_factors,
+                atomization_consistent=args.atomization_consistent,
+            )
 
             en_name, force_name = get_data_names(args.qm_method, args.basis_set)
 
@@ -221,13 +260,15 @@ def main(args):
 
             training_modules, db_info = assemble_for_training(train_loss, validation_losses)
 
-            database = load_db(db_info,
-                               en_name,
-                               force_name,
-                               n_workers=args.n_workers,
-                               seed=args.seed,
-                               anidata_location=args.anidata_location,
-                               use_ccx_subset=args.use_ccx_subset)
+            database = load_db(
+                db_info,
+                en_name,
+                force_name,
+                n_workers=args.n_workers,
+                seed=args.seed,
+                anidata_location=args.anidata_location,
+                use_ccx_subset=args.use_ccx_subset,
+            )
 
             from hippynn.pretraining import hierarchical_energy_initialization
 
@@ -237,62 +278,77 @@ def main(args):
             if args.use_ccx_subset:
                 patience *= 4
 
-            setup_params = setup_experiment(training_modules,
-                                            device=args.gpu,
-                                            batch_size=args.batch_size,
-                                            init_lr=args.init_lr,
-                                            patience=patience,
-                                            max_epochs=args.max_epochs,
-                                            stopping_key=args.stopping_key,
-                                            )
+            setup_params = setup_experiment(
+                training_modules,
+                device=args.gpu,
+                batch_size=args.batch_size,
+                init_lr=args.init_lr,
+                patience=patience,
+                max_epochs=args.max_epochs,
+                stopping_key=args.stopping_key,
+            )
 
             from hippynn.experiment import setup_and_train
 
-            setup_and_train(training_modules=training_modules,
-                            database=database,
-                            setup_params=setup_params,
-                            )
+            setup_and_train(
+                training_modules=training_modules,
+                database=database,
+                setup_params=setup_params,
+            )
 
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
 
-    parser.add_argument("--tag", type=str, default="TEST_MODEL_ANI1X", help='name for run')
-    parser.add_argument("--gpu", type=int, default=0, help='which GPU to run on')
-    parser.add_argument("--use_gpu", type=bool, default=False, help='Whether to use GPU')
+    parser.add_argument("--tag", type=str, default="TEST_MODEL_ANI1X", help="name for run")
+    parser.add_argument("--gpu", type=int, default=0, help="which GPU to run on")
+    parser.add_argument("--use_gpu", type=bool, default=False, help="Whether to use GPU")
 
-    parser.add_argument("--seed", type=int, default=0, help='random seed for init and split')
+    parser.add_argument("--seed", type=int, default=0, help="random seed for init and split")
 
     parser.add_argument("--n_interactions", type=int, default=2)
     parser.add_argument("--n_atom_layers", type=int, default=3)
     parser.add_argument("--n_features", type=int, default=128)
     parser.add_argument("--n_sensitivities", type=int, default=20)
     parser.add_argument("--cutoff_distance", type=float, default=6.5)
-    parser.add_argument("--lower_cutoff",type=float,default=0.75,
-            help="Where to initialize the shortest distance sensitivity")
-    parser.add_argument("--tensor_order",type=int,default=0)
+    parser.add_argument("--lower_cutoff", type=float, default=0.75, help="Where to initialize the shortest distance sensitivity")
+    parser.add_argument(
+        "--tensor_model",
+        type=str.upper,
+        default="HOP",
+        choices=["HOP", "TS", "NONE"],
+        help="Which tensor architecture to use.  Choices are 'HOP' for HIP-HOP-NN, "
+        "'TS' for HIP-NN-TS, and 'NONE' for vanilla HIP-NN'. "
+        "If tensor_order==0 then vanilla HIP-NN will "
+        "be used regardless.",
+    )
+    parser.add_argument("--tensor_order", type=int, default=0, help="tensor order $\ell$")
+    parser.add_argument("--tensor_factors", type=int, default=4, help="number of factors used (in HIP-HOP-NN only)")
     parser.add_argument("--atomization_consistent", type=bool, default=False)
 
-    parser.add_argument("--anidata_location", type=str, default='../../../datasets/ani1x_release/ani1x-release.h5')
-    parser.add_argument("--qm_method", type=str, default='wb97x')
-    parser.add_argument("--basis_set", type=str, default='dz')
+    parser.add_argument("--anidata_location", type=str, default="../../../datasets/ani1x_release/ani1x-release.h5")
+    parser.add_argument("--qm_method", type=str, default="wb97x")
+    parser.add_argument("--basis_set", type=str, default="dz")
 
-    parser.add_argument("--force_training", action='store_true', default=True)
+    parser.add_argument("--force_training", action="store_true", default=True)
 
-    parser.add_argument("--batch_size",type=int, default=256)
-    parser.add_argument("--init_lr",type=float, default=1e-3)
-    parser.add_argument("--patience",type=int, default=5)
-    parser.add_argument("--max_epochs",type=int, default=500)
-    parser.add_argument("--stopping_key",type=str, default="T-RMSE")
+    parser.add_argument("--batch_size", type=int, default=256)
+    parser.add_argument("--init_lr", type=float, default=1e-3)
+    parser.add_argument("--patience", type=int, default=5)
+    parser.add_argument("--max_epochs", type=int, default=500)
+    parser.add_argument("--stopping_key", type=str, default="T-RMSE")
 
-    parser.add_argument("--use_ccx_subset",type=bool, default=False,
-                        help="Train only to configurations from the ANI-1ccx subset."
-                             " Note that this will still use the energies using the `qm_method` argument."
-                             " *Note!* This argument will multiply the patience by a factor of 4.")
+    parser.add_argument(
+        "--use_ccx_subset",
+        type=bool,
+        default=False,
+        help="Train only to configurations from the ANI-1ccx subset."
+        " Note that this will still use the energies using the `qm_method` argument."
+        " *Note!* This argument will multiply the patience by a factor of 4.",
+    )
 
-
-    parser.add_argument("--noprogress", action='store_true', default=False, help='suppress progress bars')
-    parser.add_argument("--n_workers", type=int, default=2, help='workers for pytorch dataloaders')
+    parser.add_argument("--noprogress", action="store_true", default=False, help="suppress progress bars")
+    parser.add_argument("--n_workers", type=int, default=2, help="workers for pytorch dataloaders")
     args = parser.parse_args()
 
     main(args)