+++ hippynn/networks/__init__.py
@@ -2,4 +2,4 @@
 Composite pytorch modules that represent larger models.
 """
 
-from . import hipnn
+from . import hipnn, hiphop